Stored static framework offer filters in allocator framework class.
This commit unpacks static offer filters configured in `FrameworkInfo`
into a dedicated field in the allocator-specific `Framework` class which
uses `ResourceQuantities`.

When the allocator receives an update on the framework's information it
will use any newly specified offer filters. This is validated with a
follow-up patch.

We will update the allocator to make use of this information in a
subsequent patch.

Review: https://reviews.apache.org/r/69889/