Refactored 'support/verify-reviews.py' to be closer to commit 7412179.
The goal of this commit is to make 'verify-reviews.py' as close as
possible to its last known working state (commit
74121798f24fca372180b8c4bc00b4df07d46240).

The diff between 'verify-reviews.py' at this commit and 7412179 is
available here: https://www.diffchecker.com/cmfaM83O All the new
features added since 7412179 have been added again and the code
has been made Python 3 compatible again.

The goal is to have a diff as minimal as possible with improvements
regarding logs so that we do not face CI issues anymore. Changes
have been made regarding how we run commands from the script so
that they are run from the repository instead of where the
script is being called.

Review: https://reviews.apache.org/r/69817/