Added a test for MESOS-9554.
This test ensures that the framework allocation loop does not
incorrectly break in the case that a framework is incapable
of receiving the resources on an agent. In this case, the loop
should continue as other frameworks may be capable of receiving
the resources.

Review: https://reviews.apache.org/r/69942