Add resource decorator hook to implicitly allocate mandatory resources.
This commit introduces a master hook to decorate task resources in
order to allocate a given amount of custom resource if the framework
does not support it yet.

For instance, if one introduces a new custom resource in a cluster
running frameworks not supporting this resource, there will be a mixed
set of tasks consuming and not consuming this resource leading to
isolation issues. By implementing this hook, a default amount can be
allocated for a custom resources on behalf of the framework so that
every tasks end up consuming this resource and Mesos can take it into
account.

This implicit allocation of resource helps introducing a new custom
resource in the clusters because, before this patch, all frameworks
needed to be patched before introducing the new resource while now a
default value can be applied for the frameworks not supporting the
resource yet meaning the patches can be done later.

https://issues.apache.org/jira/browse/MESOS-9315

Review: https://reviews.apache.org/r/69938/