Added cleanup logic for orphaned operations.
The resources used by orphaned operations are not accounted for
in the Slave struct's `usedResources`, and must be treated differently
when cleaning up the operation.  Removal of non-terminal orphan
operations will instead augment the `totalResources`.

NOTE: The only codepath that can remove non-terminal operations is
when removing the agent or a resource provider.  In this case,
there is no need to update the allocator with the augmented
`totalResources` because the removal codepaths will already
update the allocator.

Review: https://reviews.apache.org/r/69962