Made SLRP clean up mount directories for destroyed MOUNT disks.
When the SLRP stage/publish a CSI volume, it creates the staging/target
paths, but only the target path is cleaned up during unpublish.
Moreover, even if the CSI plugin does not support node-staging, the
current cleanup is still not enough, as the parent directory of the
target path is not removed. This patch fixes this problem.

Review: https://reviews.apache.org/r/69970