Fixed verify-reviews.py to not abort review verification prematurely.
Previously `support/verify-reviews.py` would abort prematurely
whenever a review a patch could not be applied. This was due to
the `shell` function used to call `support/apply-reviews.py` invoking
`exit(1)` on the first error which stopped iteration over all
outstanding reviews.

In this patch that explicit call to `exit` is removed, and instead we
let a possible `subprocess.CalledProcessError` propagate up for it to
be handled at a higher level where it is translated into a
`ReviewError`. Currently this will post a note on the review in
question to (1) notify the submitter, and (2) prevent the review from
being checked again.

With the changes here the script will not stop verify reviews in cases
where a review does not apply.

Review: https://reviews.apache.org/r/70000