Removed operations when removing resource providers.
When a resource provider explicitly disconnects from the agent, the
agent will send a `UpdateSlaveMessage` to the master, telling the
master to remove the resource provider.  If there are any operations
associated with the resource provider, they must be removed too,
because there is no way to make forward progress on resource provider
operations without a resource provider.

This removes a potential memory leak in the master's Framework structs.

Review: https://reviews.apache.org/r/70014