Added test for terminal operation updates after master failover.
This test covers a corner case where an agent reregisters with the
master with a pending operation, but the operation's originating
framework is unknown.  This can occur in a variety of situations like:
  * the master fails over and a framework never reregisters,
  * a completed framework is rotated out of the master's memory with
    pending operations, or
  * an agent with pending operations is migrated from one cluster to
    another.

In this case, the master should "adopt" the orphan operation only
after a delay.  This gives the framework some time to reregister.
But if the framework does not reregister in time, the master will
be in charge of acknowledging operation status updates.

Review: https://reviews.apache.org/r/70040