Store `logrotate` config in memfd file instead of container's sandbox.
Previously, logrotation module stored the `logrotate` configuration
file in container's sandbox directory, so that it was garbage collected
together with the container's sandbox. If the container's task had
permissions to modify this configuration file, it was possible to run
any command under an unprivileged user. This patch stores `logrotate`
config in an nonymous temporary file via `memfd`, so logrotation module
can pass a path to procfs instead of container's sandbox. This approach
solves the aforementioned security issue on Linux when the
`ENABLE_LAUNCHER_SEALING` configuration flag is specified.

Review: https://reviews.apache.org/r/70010/