Changed the definition of `uid_t` and `gid_t` to UNIT for Windows.
This is to be consistent with other platforms. And we also need this
change for https://reviews.apache.org/r/69479/, otherwise that patch
cannot be compiled on Windows because we do not have a matched `parse`
function for `int` in 3rdparty/stout/include/stout/flags/parse.hpp.

Review: https://reviews.apache.org/r/69553/