Avoid dereferencing removed executors and launching containers for them.
When launching executors and tasks, there is no guarantee that the
executors still remain after `Slave::publishResources` is returned. If
not, the executor struct should not be dereferenced and the executor
containers should not be launched at all.

NOTE: The patch makes `Slave::launchExecutor` called asynchronously even
if there is no secret generator. However this should not affect the
correctness of executor launching.

Review: https://reviews.apache.org/r/70084