Removed `support/cpplint.patch`.
This change removes explicit tracking of a patch against upstream
cpplint in the tree and replaces it with a clear reference against the
upstream repository and revision used in the modified file.

The patch was never used by us against any upstream version, and was
not e.g., a legal requirement. The process around creating the patch
was cumbersome and error prone. Removing the patch from the tree
should not only make it easier to evolve our modifications going
forward, but also remove a lot of noise from changes.

Review: https://reviews.apache.org/r/70098