Removed outdated scaling framework.
This old code was never updated for recent changes, is not documented,
and currently does not work anymore. Removing it for now. Should we
decide to bring it back in the future we should make sure it is tested
automatically so it does not go out of sync again, and also that it is
documented and discoverable by users.

Review: https://reviews.apache.org/r/70103