Added a new registry field `minimum_capabilities`.
This patch adds a new registry field `minimum_capabilities`
to faciliatate safe downgrade of masters.  This is a list of strings
which will be written whenever the master uses a feature that may
be backwards incompatible.  If an older master sees an unknown feature
in the list, the older master will refuse to recover and try to provide
remediation steps instead.

Also see: MESOS-8878.

Review: https://reviews.apache.org/r/67761/