Added minimum capability check during master recovery.
Upon recovery, the master will compare its own capabilities against the
list of strings read from the registry.  If the master is missing any
of the capabilities read from the registry, the master will refuse to
recover and try to provide remediation steps instead.

Modifications to the registry's list will be added in future, within
the logic for other RegistryOperations.  For example, if support for
Quota limit bursting (MESOS-8068) is added to a future master,
setting a quota limit should cause the UpdateQuota RegistryOperation
to add the appropriate minimum capability.  A RemoveQuota
RegistryOperation would remove the minimum capability if there are no
other quota limits in use.

Also adds a dedicated test.

Review: https://reviews.apache.org/r/67762/