Added ARCHIVE_EXTRACT_SECURE_NODOTDOT flag to archiver default.
This enables a security flag provided by libarchive, which disallows
extraction of archives that contain '..' in hardlinks or files.
Without this flag, it is possible to provide the archiver with
an archive and overwrite arbitrary files in the user's parent directory
or further up.

Review: https://reviews.apache.org/r/70142/