Updated quota capacity check to not exclude disconnected agents.
The exclusion of disconnected agents produces a rather unstable
calculated capacity in the face of upgrades, transient network
issues, etc. To provide a more stable calculated capacity, this
includes registered agents, even when they are disconnected.

This also adds a TODO to include recovered agents so that the
calculated capacity after a failover is the same as it was prior
to the failover.

Review: https://reviews.apache.org/r/70274