Tracked operation IDs in the agent.
This patch updates the agent to track operation IDs in a
hashmap. This map will be used to satisfy some reconciliation
requests which are forwarded from the master, when the agent
must be able to find the internal operation UUID for an
operation which the master did not recognize.

Review: https://reviews.apache.org/r/70242/