Improved operation reconciliation for unsubscribed resource providers.
This patch updates the operation reconciliation pipeline to
forward some framework-initiated reconciliation requests to
the agent. In cases where an explicit reconciliation request
specifies an operation which is not recognized on a resource
provider which is also not recognized, the master will
forward the request to the agent so that the resource
provider manager can satisfy the request based on whether or
not the resource provider has been seen before.

Review: https://reviews.apache.org/r/70243/