Bundled CSI spec 1.1.0.
Since the CSI v1 spec proto file depends on certain proto files in the
Protobuf library, we have to ensure the Protobuf library's include path
is in the proto paths of the `protoc` command when compiling the CSI
spec proto file. Specifically in Autotools, this path is passed through
the `PROTOBUF_PROTOCFLAGS` variable when building with an unbundled
protobuf library.

Review: https://reviews.apache.org/r/70360