Mounted /proc properly a container shares pid namespace with its parent.
If a container shares the same pid namespace as its parent and is not a
top level container. It might or might not share the same pid namespace
as the agent. In this case, we need to re-mount `/proc`.

One caveat here is that: in the case where this container does share the
pid namespace of the agent (because its parent shares the same pid
namespace of the agent), mounting `/proc` at the same place will result
in EBUSY.

As a result, we need to "move" (MS_MOVE) the mounts under `/proc` to a
new location and mount the `/proc` again at the old location.

See MESOS-9529 for details.

Review: https://reviews.apache.org/r/70356
(cherry picked from commit f88880b6a5568cc5573be8f0ac5f9dcdd48914b5)