Initialized resource provider manager earlier when recovering.
When recovering and reusing the same agent ID the resource provider
manager can be initialized before e.g., recovering executors. This patch
move the initialization to such an earlier point. This e.g., allows to
successfully publish resources via the manager when HTTP-based executors
resubscribe which previously ran into an assertion failure.

If the agent ID is not reused we still need to wait for the agent to
register with the master which would assign an agent ID. In that case we
do not expect any executors to resubscribe.

Review: https://reviews.apache.org/r/70368/