Added a test to ensure correct headroom accounting.
This test verifies that the headroom accounting is correct in the
presence of subrole allocation. In particular, we need to ensure
that a subrole's allocation is accounted only once (not multiple
times in itself as well as all of its ancestors) when calculating
available the quota headroom.

Review: https://reviews.apache.org/r/70390