Enhanced HierarchicalAllocatorTest.QuotaWithNestedRoleReservation.
This test missed an issue in https://reviews.apache.org/r/70393/
because it was only checking that the quota consumption for "a"
was >= the subrole "a/b" reservation. Updated this test to ensure
that the quota consumption is exactly equal to the subrole "a/b"
reservation.

Review: https://reviews.apache.org/r/70395