Enabled launcher sealing for RPM packages.
We enable this flag since with it disabled certain public functions
are not available making it hard to e.g., write modules against this
version of Mesos.

While launcher sealing depends on a recent kernel, the platform we
build RPMs for already satisfies the requirements.

Review: https://reviews.apache.org/r/70295