Prevented bundled glog cmake build from finding system gflags.
Since glog-0.4.0 the glog cmake build seems to find a system-wide
installation of gflags and use that to implicitly enable additional
features which we do not use. Due to the way glog is integrated into
our cmake build we can then end up in a situation where we
successfully build the bundled glog, but cannot include all its
required headers.

This patch fixes the cmake flags of the bundled glog so it never uses
gflags which 1) is more deterministic and 2) addresses above mentioned
build issue.

Review: https://reviews.apache.org/r/70466/