Supported docker manifest v2s2 config with image GC.
After the docker manifest v2 schema2 support, the manifest config
is saved as a layer in image store. However, the config file
cannot survive if the image GC is triggered.

This patch supports checkpointing the config path in provisioner,
so that the provisioner has the information of configs so that
they will not be deleted during GC.

(cherry picked from commit cd75916020080232a27df4597f9de4dc2dc961d9)