Fixed the flaky ExamplesTest.DynamicReservationFramework.
The test failed in MESOS-5804 due to the following race:

  1. Framework launches task T, moves from RESERVED to
     TASK_RUNNING state.
  2. Allocation cycle triggers and will send the unreserved
     resources to the framework.
  3. Before the offer gets to the framework, task T finishes and
     framework moves from TASK_RUNNING to RESERVED.
  4. In the RESERVED state, the framework expects the reservation
     in the offer. But, it's coming in a later offer, and the one
     that arrives is for the unreserved resources since it was
     generated while the task was still running.

Tne fix applied here for this specific race is to use a 2 week
filter rather than a 0 second filter. That would ensure that the
unreserved resources do not get re-offered to the framework on
their own. However, this fix does not work until MESOS-9616 is
resolved.

Review: https://reviews.apache.org/r/70508