Added authorization for `UpdateQuota` call in the master.
A new authorizable action `UPDATE_QUOTA_WITH_CONFIG` is added.
This disambiguates with the old action `UPDATE_QUOTA` which
are used for the old `SetQuota` and `RemoveQuota` calls.
`UPDATE_QUOTA` action requires `QuotaInfo` as the object while
the new `UpdatedQuota` call sets the `value` field of the object
using the role associated with the quota. To keep it compatible
with any external authorization modules, a new action is introduced.

Review: https://reviews.apache.org/r/70549