Relaxed protobuf union validation strictness.
As part of MESOS-6874, the master validates protobuf unions passed as
part of an ExecutorInfo::ContainerInfo.  This prevents a task from
specifying, for example, a ContainerInfo::MESOS, but filling
out the docker field (which is then ignored by the agent).

This validation change is actually an API change, because previously
runnable ExecutorInfo's and TaskInfo's will now fail validation.
This has two visible effects on clusters:
  * Agents running containers with invalid protobuf unions will not
    be able to reregister with the master.
  * Existing frameworks will not be able to re-launch the same tasks
    that were working before a Mesos master upgrade.

This changes the validation to print a warning instead.  Where possible,
the warning will provide some information to indicate which task
or executor is sending the invalid protobuf.

Review: https://reviews.apache.org/r/70546/