Logged headroom related info in the allocator.
This patch logs `requiredHeadroom` and `availableHeadroom`
before each allocation cycle and logs `requiredHeadroom`,
resources and number of agents held back for quota `headroom`
as well as resources filtered at the end of each allocation
cycle.

While we also held resources back for quota headroom in the
first stage, we do not log it there. This is because in the
second stage, we try to allocate all resources (including the
ones held back in the first stage). Thus only resources held
back in the second stage are truly held back for the whole
allocation cycle.

Review: https://reviews.apache.org/r/70570