Randomized the agents in the second allocation stage.
Before this patch, agents are randomized before the 1st
allocation stage (the quota allocation stage) but not in
the 2nd stage. One perceived issue is that resources on
the agents in the front of the queue are likely to be mostly
allocated in the 1st stage, leaving only slices of resources
available for the second stage. Thus we may see consistently
low quality offers for role/frameworks that get allocated first
in the 2nd stage.

This patch randomizes the agents again before the 2nd stage to
to "spread out" the effect of the 1st stage allocation.

Review: https://reviews.apache.org/r/70613