Changed Agent GET_STATE for completed executor's tasks (Part 2).
This is another code path where an executor can "complete" with
a non-terminal task (based on last status update).  This happens when
the Framework or Master TEARDOWN calls are made, for a framework.
If the executor does not send a TASK_KILLED in time, the agent will
still mark the executor as complete and kill it.

Unlike the other code path, this one does not require an agent restart.
The agent in this state will return a GET_STATE response where
non-terminal tasks of a completed executor (of a completed framework)
will appear under `launched_tasks`.  This may provide misleading
information about the total number of tasks running.

This commit adds extra logic to place these non-terminal tasks under
the `terminated_tasks` category, and adds a regression test.

Review: https://reviews.apache.org/r/70641