Introduced a function for validating a `FrameworkInfo` update.
This patch moves the code validating a new `FrameworkInfo` against
the current one into a separate function. Note that the function
currently *does not* check `user` and `checkpoint`, and that this
will be done in a subsequent patch as some refactoring is needed
to make that possible.

Review: https://reviews.apache.org/r/70666/