Moved the logic of broadcating framework updates into a separate method.
This patch consolidates the scattered logic of sending framework updates
(namely, the FRAMEWORK_ADDED event to master V1 API subscribers and
UpdateFrameworkMessage to agents).

NOTE: The UpdateFrameworkMessage is now also being sent in the case of
non-`forced` subscription request from a framework which has already
been registered. Previously it was not being sent in this case -
this likely is a bug.

Review: https://reviews.apache.org/r/70665/