Added `struct Role` to track role reservations and framework IDs.
This patch introduces `struct Role` which contains the framework IDs
and reservatons tracked under a role. Two separate hashmaps used for
this purpose previously are replaced by a single hashmap, thus the need
to keep the two consistent with each other is removed. This simplifies
the role tracking logic in the allocator. The patch introduces minimal
to no performance impact.

Review: https://reviews.apache.org/r/70591/