Add containerizer support for masking paths.
Add support to the `filesystem/linux` isolator for masking container
paths. Add a set of standard default paths to be masked, as derived
from commonly used container runtimes. These paths either expose
information about other system processes, or capabilities that
should not be exposed to untrusted containers.

We don't mask if the container is privileged, which is defined
as sharing the host's PID namespace. For nested containers, we
verify that the PID namespace is shared from the host all the way
up the tree.

Review: https://reviews.apache.org/r/70678/