Decoupled quota limits from quota guarantees in the allocator.
Currently, resource quota acts as both guarantees and limits.
This patch adds a field `quotaLimits` in the role struct.
This paves the way for enforcing guarantees and limits
separately in the allocator.

Also added a helper `getLimits` to look up quota limits of a
given role.

Review: https://reviews.apache.org/r/70716