Windows: Removed multiple categories of sources from the build.
This removes some categories of sources from the Windows build,
where it is possible to do so with minimal ifdef-ing.

The features removed are all Linux-specific features that cannot be
feasibly ported to Windows, including:
  * Container Storage Interface (CSI)
  * Docker image provisioning, specifically related to V2
  * Open Container Interface
  * Volume GID Manager

Protobufs are excluded where possible, but many of the above categories
of protobufs are interleaved with other protobufs or source code,
which makes exclusion non-trivial.  For example, libseccomp is a
Linux-only feature, but its protobufs are now required to build the
Mesos containerizer's protobufs.

Docker image provisioning was semi-trivial to exclude, because the
related components (provisioner & URI fetcher) are somewhat modularized.

Review: https://reviews.apache.org/r/70740