Fixed incorrect comparision introduced in `0ab27c674b7`.
As written the code was comparing a `TaskState` against a
`v1::TaskState` which are different enums. This was not only brittle and
not guaranteed to work, but very likely also not intended.

This patch fixes the code to compare values of the same enum type.

Review: https://reviews.apache.org/r/70721