Allow roles to burst up to quota limits in the allocator.
This patch allows roles to burst above their quota guarantees
up to the quota limits.

In the first allocation stage which is dedicated to roles
with non-default guarantees, we allow these roles to burst
up to their limits for resources that they have guarantees
set as well as for other resources allocated together
with the guaranteed resources. These bursts are subject to
global headroom enforcement.

In the second allocation stage, we allow all roles to burst
up to their limits while maintaining the global headroom.
In this stage, if some unreserved non-revocable scalar resources
on the agent need to be held back to maintain the headroom,
we will "chop" them to make the largest offer possible without
breaking the quota headroom. Previously, if any resources are
needed on this agent to maintain the headroom, NO unreserved
scalar resources will be offered at all.

Also fixed an affected test.

Review: https://reviews.apache.org/r/70738