Fixed test `QuotaRoleAllocateNonQuotaResource`.
The test was failing because:

After `agent3` is added, it misses a settle call where the allocation
of `agent3` is racy.

In addition, after MESOS-8456, the allocator now offers non-quota
resources on an agent (even that means "chopping") on top of a role's
satisfied guarantees instead of skipping the agent all-together.

This patch fixes the test by expecting the right amount of resources
to be allocated from `agent3`.

Review: https://reviews.apache.org/r/70824