Updated webui Roles tab to consistently display '-' for 0 entries.
The Roles tab currently uses an inconsistent mix of '0', '0 B' and '-'
for displaying the allocation, guarantee, and limit. Since it's easier
to scan the table for non-empty entries when '-' is used, this updates
the every empty case to use '-'.

We may also want to consider using the empty string '' in the future,
as that may help make the table even easier to scan visually than '-'.

Review: https://reviews.apache.org/r/70879