Assign cgroup processes after configuring the subsystem.
Currently, the PID targeted by the cgroups isolator is moved into
the cgroup before the subsystem runs to apply any type-specific
cgroup configuration. We should reverse the order of this so that
the PID is only moved once the cgroup is fully configured by the
subsystem.

A specific case that can happen is where a PID is assigned to a net_cls
cgroup before that cgroup has its class ID set.  This intermediate
process state can be observed by system monitoring process, causing
confusion that is hard to debug.

Review: https://reviews.apache.org/r/70863/