Made `/roles` endpoint also return quota limits.
Now that guarantees are decoupled from limits, we should
return limits and guarantees separately in the `/roles` endpoint.

Three incompatible changes are introduced:

- The `principal` field is removed. This legacy field was used to
record the principal of the operator who configured the quota.
So that later, if a different operator with a different principal
wants to modify the quota, the action can be properly authorized.
This use case has since been deprecated and the principal field
will no longer be filled going forward.

- Resources with zero quantity will no longer be included in
the `guarantee` field.

- The `guarantee` field will continue to be filled.
However, since we are decoupling the quota guarantee from the limit.
One can no longer assume that the limit will be the same as guarantee.
A separate `limit` field is introduced.

Before, the response might contain:
```
{
  "quota": {
    "guarantee": {
      "cpus": 1,
      "disk": 0,
      "gpus": 0,
      "mem": 512
    },
    "principal": "test-principal",
    "role": "foo"
  }
}
```

After:
```
{
  "quota": {
    "guarantee": {
      "cpus": 1,
      "mem": 512
    },
    "limit": {
      "cpus": 1,
      "mem": 512
    },
    "role": "foo"
  }
}
```

Also fixed an affected test.

Review: https://reviews.apache.org/r/70915