Added master minimum capability `QUOTA_V2`.
This adds a new enum for the revamped quota feature
in the master. When quota is configured in Mesos 1.9
or higher, the quota configurations will be persisted
into the `quota_configs` field in the registry. And
the `QUOTA_V2` minimum capability will be added to the
registry as well. This will prevent any master downgrades
until `quota_configs` becomes empty. This can be done by
setting the quota of the roles listed in `quota_configs`
back to the default (no guarantees and no limits).

Note, since at the moment of adding this patch, the master
is not yet capable of handling the new quota API. The
`capability` is not added to the `MASTER_CAPABILITIES`.
That should be done later together with the patches that
enables master for handling the new quota calls.

Review: https://reviews.apache.org/r/70949